+++ lib/numeric.dart
@@ -1,11 +1,13 @@
-/// Form fields related digit-only or numeric values.
+/// Form fields related to numeric values.
 ///
 /// Examples of numeric input data:
 ///
-/// - a three-digit numeric code
-/// - a six-digit password
-/// - a hexadecimal input
+/// - a three-digit numeric code.
+/// - a six-digit password.
+/// - a hexadecimal input.
 /// - Minimum Order Quantity — the fewest number of units required at one time.
 library numeric;
 
 export 'src/numeric/digit_field.dart';
+export 'src/numeric/int_field.dart';
+export 'src/numeric/num_field.dart';